Renamed puzzle folder, improved UI, added font, fixed AI pathfinding
