--- Assets/Popup Ui.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6063541d24a96b648907e0946af9bf33
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
