--- Assets/Popup Ui/Prefabs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: dcd910551f10dc143af0280709824c99
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
