--- Assets/Popup Ui/Prefabs/HintUI.uxml
@@ -1,10 +1,0 @@
-<ui:UXML xmlns:ui="UnityEngine.UIElements" xmlns:uie="UnityEditor.UIElements" xsi="http://www.w3.org/2001/XMLSchema-instance" engine="UnityEngine.UIElements" editor="UnityEditor.UIElements" noNamespaceSchemaLocation="../../../UIElementsSchema/UIElements.xsd" editor-extension-mode="False">
-    <ui:GroupBox>
-        <ui:GroupBox name="HintDialogue" style="background-color: rgb(56, 45, 35); border-top-left-radius: 4px; border-bottom-left-radius: 4px; border-top-right-radius: 4px; border-bottom-right-radius: 4px; justify-content: space-around; overflow: hidden; position: absolute; top: 122px; right: 186px; padding-left: 3px; padding-right: 3px; padding-top: 3px; padding-bottom: 3px; visibility: hidden; max-width: 35%; margin-left: 0; margin-right: 0; margin-top: 0; margin-bottom: 0;">
-            <ui:VisualElement style="background-color: rgb(121, 87, 63); width: 100%; height: 100%; margin-left: 0; margin-right: 0; margin-top: 0; margin-bottom: 0; border-top-left-radius: 4px; border-bottom-left-radius: 4px; border-top-right-radius: 4px; border-bottom-right-radius: 4px; border-left-width: 2px; border-right-width: 2px; border-top-width: 2px; border-bottom-width: 2px; border-left-color: rgb(200, 173, 146); border-right-color: rgb(200, 173, 146); border-top-color: rgb(200, 173, 146); border-bottom-color: rgb(200, 173, 146); justify-content: space-around; flex-wrap: wrap; overflow: hidden; padding-left: 3px; padding-right: 3px; padding-top: 3px; padding-bottom: 3px;">
-                <ui:Label text="Label" display-tooltip-when-elided="true" name="HintText" style="flex-wrap: nowrap; overflow: hidden; flex-basis: auto; align-items: stretch; width: auto; height: auto; white-space: normal;" />
-                <ui:Button text="OK" display-tooltip-when-elided="true" name="HintOkButton" style="width: 48px; margin-left: auto; margin-right: auto; overflow: hidden; display: none;" />
-            </ui:VisualElement>
-        </ui:GroupBox>
-    </ui:GroupBox>
-</ui:UXML>
