--- Assets/Popup Ui/Prefabs/HintUI.uxml.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: f4a81dce15d12fe41822811d91a52a54
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 13804, guid: 0000000000000000e000000000000000, type: 0}
