--- Assets/Popup Ui/Prefabs/Popup.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ba1b1f0728ecfe34d95dd46a8da26994
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
