--- Assets/Popup Ui/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3eb6d946388d4b54fb0733863b6d7bde
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
