--- Assets/Popup Ui/Scripts/CalculatorScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 24c82a33e7321364c9e865bde53a7897
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
