--- Assets/Popup Ui/Scripts/HintDialogue.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 47880af37db7f7a47bb47a4f9511bf08
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
