--- Assets/Popup Ui/Scripts/InformationPanelScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 68e572d2b8693d246a75a8a84bcabc42
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
