--- Assets/Popup Ui/Scripts/PopupUi.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 533953a8028cdba468051159f2f7cf68
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
