--- Assets/Popup Ui/Scripts/SimonsaysScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 46a2af4890eafcb41b98abb13c293b89
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
