+++ Assets/Prefabs/InteractableCircle.prefab
@@ -51,8 +51,8 @@ MonoBehaviour:
     m_PersistentCalls:
       m_Calls:
       - m_Target: {fileID: 0}
-        m_TargetAssemblyTypeName: DoorController, Assembly-CSharp
-        m_MethodName: ToggleDoor
+        m_TargetAssemblyTypeName: PuzzelController, Assembly-CSharp
+        m_MethodName: OpenPuzzel
         m_Mode: 1
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
