+++ Assets/Scripts/Interfaces/Interactable.cs
@@ -39,11 +39,6 @@ public class Interactable : MonoBehaviour
         }
     }
 
-    void Start()
-    {
-        
-    }
-
     // Update is called once per frame
     void Update()
     {
