+++ Assets/UI/Fonts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c748d7cd926bdf2448414fe3aa6114f6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
