+++ Assets/UI/Fonts/Egyptaxt SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c8099edd7da67c84bbb61f2e65ee65f1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
