+++ Assets/UI/UI Screens/mainMenu.uxml
@@ -1,11 +1,11 @@
 <ui:UXML xmlns:ui="UnityEngine.UIElements" xmlns:uie="UnityEditor.UIElements" editor-extension-mode="False">
-    <ui:IMGUIContainer style="height: 656px; background-image: none; align-items: center; -unity-background-scale-mode: stretch-to-fill; margin-top: 52px;">
-        <ui:VisualElement style="background-image: url(&apos;project://database/Assets/UI%20Toolkit/Misc/Cartoon%20RPG%20UI_Window%20-%20Dialogue.png?fileID=2800000&amp;guid=79c371e0ece1ab741a63dd09faa402b0&amp;type=3#Cartoon RPG UI_Window - Dialogue&apos;); width: 545px; height: 87px; justify-content: center; align-items: center;">
-            <ui:Label text="Pyramid Quest" display-tooltip-when-elided="true" style="height: 56px; width: 482px; -unity-text-align: middle-center; -unity-font: resource(&apos;FlatSkin/Font/Roboto-Medium&apos;); font-size: 24px;" />
+    <ui:IMGUIContainer style="height: 1022px; background-image: none; align-items: center; -unity-background-scale-mode: stretch-to-fill; margin-top: 52px; justify-content: center;">
+        <ui:VisualElement style="background-image: url(&apos;project://database/Assets/UI/UI%20Toolkit/Misc/Cartoon%20RPG%20UI_Window%20-%20Dialogue.png?fileID=2800000&amp;guid=79c371e0ece1ab741a63dd09faa402b0&amp;type=3#Cartoon RPG UI_Window - Dialogue&apos;); width: 545px; height: 87px; justify-content: center; align-items: center;">
+            <ui:Label text="Pyramid Quest" display-tooltip-when-elided="true" style="height: 56px; width: 482px; -unity-text-align: middle-center; -unity-font: resource(&apos;FlatSkin/Font/Roboto-Bold&apos;); font-size: 32px; color: rgb(255, 255, 255); -unity-font-definition: url(&apos;project://database/Assets/UI/Fonts/Egyptaxt%20SDF.asset?fileID=11400000&amp;guid=c8099edd7da67c84bbb61f2e65ee65f1&amp;type=2#Egyptaxt SDF&apos;);" />
         </ui:VisualElement>
         <ui:VisualElement name="buttonContainer" style="height: 280px; flex-wrap: nowrap; justify-content: space-around; align-items: center;">
-            <ui:Button display-tooltip-when-elided="true" name="play" style="background-image: url(&apos;project://database/Assets/UI%20Toolkit/Buttons/Text/Button%20Text%20-%20Play.png?fileID=2800000&amp;guid=c5f91d73f1ebba1478c7016c811d9e52&amp;type=3#Button Text - Play&apos;); -unity-background-scale-mode: scale-and-crop; height: 100px; width: 289px;" />
-            <ui:Button display-tooltip-when-elided="true" name="exit" style="background-image: url(&apos;project://database/Assets/UI%20Toolkit/Buttons/Text/Button%20Text%20-%20Exit.png?fileID=2800000&amp;guid=87d6022f84cbac344befae7a39f2a5e2&amp;type=3#Button Text - Exit&apos;); height: 100px; width: 284px; -unity-background-scale-mode: scale-and-crop;" />
+            <ui:Button display-tooltip-when-elided="true" name="play" style="background-image: url(&apos;project://database/Assets/UI/UI%20Toolkit/Buttons/Text/Button%20Text%20-%20Play.png?fileID=2800000&amp;guid=c5f91d73f1ebba1478c7016c811d9e52&amp;type=3#Button Text - Play&apos;); -unity-background-scale-mode: scale-and-crop; height: 100px; width: 289px;" />
+            <ui:Button display-tooltip-when-elided="true" name="exit" style="background-image: url(&apos;project://database/Assets/UI/UI%20Toolkit/Buttons/Text/Button%20Text%20-%20Exit.png?fileID=2800000&amp;guid=87d6022f84cbac344befae7a39f2a5e2&amp;type=3#Button Text - Exit&apos;); height: 100px; width: 284px; -unity-background-scale-mode: scale-and-crop;" />
         </ui:VisualElement>
     </ui:IMGUIContainer>
 </ui:UXML>
